crm/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
gunicorn
psycopg2-binary
whitenoise
dj-database-url
